+++ docs/source/minification_options/prefer_single_line.py
@@ -1,0 +1,5 @@
+import os
+import sys
+
+name = "world"
+print("Hello, " + name)