+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and will output source code compatible with the version of the interpreter it is
 This means that if you minify code written for Python 3.11 using python-minifier running with Python 3.12,
 the minified code may only run with Python 3.12.
 
+## [3.2.0] - 2025-12-31
+
+### Added
+- New `--prefer-single-line` option to use semicolons instead of newlines between top-level statements when there is no difference in output size. This doesn't make the output any smaller, but may be preferred.
+- Constant folding can now fold unary operators (e.g. `-1`, `not True`, `~0`) in addition to binary operators.
+
 ## [3.1.1] - 2025-12-11
 
 ### Fixed
@@ -311,6 +317,7 @@ the minified code may only run with Python 3.12.
 - python-minifier package
 - pyminify command
 
+[3.2.0]: https://github.com/dflook/python-minifier/compare/3.1.1...3.2.0
 [3.1.1]: https://github.com/dflook/python-minifier/compare/3.1.0...3.1.1
 [3.1.0]: https://github.com/dflook/python-minifier/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/dflook/python-minifier/compare/2.11.3...3.0.0